extract: logic from ReviewStep
All the logic has been extracted into its context store.
The component has been divided into constituent subcomponents -
heading, review template content.

The store has been placed above the template base modal.
